+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/parser2/DirectEventParser.kt
@@ -21,21 +21,21 @@ import com.squareup.moshi.JsonReader
 import com.squareup.moshi.Moshi
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.parser2.adapters.DateAdapter
-import io.getstream.chat.android.client.parser2.event.AttachmentAdapter
-import io.getstream.chat.android.client.parser2.event.ChannelInfoAdapter
-import io.getstream.chat.android.client.parser2.event.DeviceAdapter
-import io.getstream.chat.android.client.parser2.event.LocationAdapter
-import io.getstream.chat.android.client.parser2.event.MessageAdapter
-import io.getstream.chat.android.client.parser2.event.MessageModerationDetailsAdapter
-import io.getstream.chat.android.client.parser2.event.MessageReminderInfoAdapter
-import io.getstream.chat.android.client.parser2.event.ModerationAdapter
-import io.getstream.chat.android.client.parser2.event.NewMessageEventAdapter
-import io.getstream.chat.android.client.parser2.event.OptionAdapter
-import io.getstream.chat.android.client.parser2.event.PollAdapter
-import io.getstream.chat.android.client.parser2.event.PrivacySettingsAdapter
-import io.getstream.chat.android.client.parser2.event.ReactionAdapter
-import io.getstream.chat.android.client.parser2.event.ReactionGroupAdapter
-import io.getstream.chat.android.client.parser2.event.UserAdapter
+import io.getstream.chat.android.client.parser2.direct.AttachmentAdapter
+import io.getstream.chat.android.client.parser2.direct.ChannelInfoAdapter
+import io.getstream.chat.android.client.parser2.direct.DeviceAdapter
+import io.getstream.chat.android.client.parser2.direct.LocationAdapter
+import io.getstream.chat.android.client.parser2.direct.MessageAdapter
+import io.getstream.chat.android.client.parser2.direct.MessageModerationDetailsAdapter
+import io.getstream.chat.android.client.parser2.direct.MessageReminderInfoAdapter
+import io.getstream.chat.android.client.parser2.direct.ModerationAdapter
+import io.getstream.chat.android.client.parser2.direct.NewMessageEventAdapter
+import io.getstream.chat.android.client.parser2.direct.OptionAdapter
+import io.getstream.chat.android.client.parser2.direct.PollAdapter
+import io.getstream.chat.android.client.parser2.direct.PrivacySettingsAdapter
+import io.getstream.chat.android.client.parser2.direct.ReactionAdapter
+import io.getstream.chat.android.client.parser2.direct.ReactionGroupAdapter
+import io.getstream.chat.android.client.parser2.direct.UserAdapter
 import io.getstream.chat.android.models.EventType
 import io.getstream.chat.android.models.MessageTransformer
 import io.getstream.chat.android.models.UserId
