+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/PollParsingTest.kt
@@ -21,11 +21,11 @@ import com.squareup.moshi.Moshi
 import io.getstream.chat.android.client.api2.mapping.DomainMapping
 import io.getstream.chat.android.client.api2.model.dto.DownstreamPollDto
 import io.getstream.chat.android.client.parser2.adapters.DateAdapter
-import io.getstream.chat.android.client.parser2.event.DeviceAdapter
-import io.getstream.chat.android.client.parser2.event.OptionAdapter
-import io.getstream.chat.android.client.parser2.event.PollAdapter
-import io.getstream.chat.android.client.parser2.event.PrivacySettingsAdapter
-import io.getstream.chat.android.client.parser2.event.UserAdapter
+import io.getstream.chat.android.client.parser2.direct.DeviceAdapter
+import io.getstream.chat.android.client.parser2.direct.OptionAdapter
+import io.getstream.chat.android.client.parser2.direct.PollAdapter
+import io.getstream.chat.android.client.parser2.direct.PrivacySettingsAdapter
+import io.getstream.chat.android.client.parser2.direct.UserAdapter
 import io.getstream.chat.android.client.parser2.testdata.PollTestData
 import io.getstream.chat.android.models.NoOpChannelTransformer
 import io.getstream.chat.android.models.NoOpMessageTransformer
